+++ packages/transaction-controller/jest.config.js
@@ -17,10 +17,10 @@ module.exports = merge(baseConfig, {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 94.31,
-      functions: 98.52,
-      lines: 99,
-      statements: 99.01,
+      branches: 94.1,
+      functions: 98.56,
+      lines: 99.03,
+      statements: 99.04,
     },
   },